requirements.txt
# Production dependencies with exact versions from pip freeze
# Excel Upload Service

# Core frameworks
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.0

# Firebase
firebase-admin==6.2.4
google-cloud-firestore==2.22.0
google-cloud-storage==3.7.0
grpcio==1.76.0
grpcio-status==1.76.0

# Excel processing
openpyxl==3.1.5
pandas==2.3.3
numpy==2.2.6

# String matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.3
Levenshtein==0.27.3
RapidFuzz==3.14.3

# Authentication
PyJWT==2.10.1

# AI (Groq)
groq==1.0.0
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
h2==4.3.0
anyio==4.12.0
sniffio==1.3.1

# Environment variables
python-dotenv==1.2.1

# Production server
gunicorn==21.2.0

# Supporting packages
requests==2.32.5
urllib3==2.6.2
certifi==2025.4.24
charset-normalizer==3.4.4
idna==3.11
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.3
six==1.17.0
